Move cloud.py.py back to cloud.py
